fix: prevenindo nao envio de email por baixa performance de servidores da ufpa
