+++ backend/meuegresso/src/main/java/labes/facomp/ufpa/br/meuegresso/controller/administrador/mensagem/MensagemAdmController.java
@@ -23,8 +23,9 @@ import labes.facomp.ufpa.br.meuegresso.dto.mensagem.MensagemDTO;
 import labes.facomp.ufpa.br.meuegresso.dto.mensagem.MensagemStatusDTO;
 import labes.facomp.ufpa.br.meuegresso.enumeration.ResponseType;
 import labes.facomp.ufpa.br.meuegresso.model.MensagemModel;
+import labes.facomp.ufpa.br.meuegresso.service.agendamento.AgendamentoService;
+import labes.facomp.ufpa.br.meuegresso.service.agendamento.impl.AgendamentoServiceImpl;
 import labes.facomp.ufpa.br.meuegresso.service.mail.MailService;
-import labes.facomp.ufpa.br.meuegresso.service.mail.impl.MailServiceImpl;
 import lombok.RequiredArgsConstructor;
 
 /**
@@ -41,7 +42,9 @@ public class MensagemAdmController {
 
 	private final MailService mailService;
 
-	private final MailServiceImpl mailServiceImpl;
+	private final AgendamentoService agendamentoService;
+
+	private final AgendamentoServiceImpl agendamentoServiceImpl;
 
 	private final ModelMapper mapper;
 
@@ -57,7 +60,7 @@ public class MensagemAdmController {
 	@GetMapping
 	@PreAuthorize(value = "hasRole('ADMIN') or hasRole('SECRETARIO')")
 	public MensagemStatusDTO consultarStatusMensagens() {
-		return mailService.getMensagensStatus();
+		return agendamentoService.getMensagensStatus();
 	}
 
 	/**
@@ -73,7 +76,7 @@ public class MensagemAdmController {
 	@GetMapping(value = "/tarefasMensagens")
 	@PreAuthorize(value = "hasRole('ADMIN') or hasRole('SECRETARIO')")
 	public Map<String, ScheduledFuture<?>> consultarTarefasMensagens() {
-		return mailService.getTasks();
+		return 	agendamentoService.getTasks();
 	}
 
 	/**
@@ -105,11 +108,7 @@ public class MensagemAdmController {
 	@ResponseStatus(code = HttpStatus.CREATED)
 	@PreAuthorize("hasRole('ADMIN') or hasRole('SECRETARIO')")
 	public String atualizarMensagem(@RequestBody @Valid MensagemDTO mensagemDTO) {
-		MensagemModel mensagemModel = mapper.map(mensagemDTO, MensagemModel.class);
-		mailService.removeScheduledTask(mensagemModel);
-		mensagemModel = mailService.update(mensagemModel);
-		mailService.setScheduleATask(mailServiceImpl, mensagemModel.getDataEnvio(), mensagemModel.getFrequente(),
-				mensagemModel.getAnual(), mensagemModel);
+		mailService.update(mapper.map(mensagemDTO, MensagemModel.class));
 		return ResponseType.SUCCESS_UPDATE.getMessage();
 	}
 
@@ -125,9 +124,10 @@ public class MensagemAdmController {
 	@PreAuthorize("hasRole('ADMIN')")
 	@ResponseStatus(code = HttpStatus.OK)
 	public String deleteById(@PathVariable(name = "id") Integer id) {
-		MensagemModel mensagemModel = mailService.findbyId(id);
-		mailService.removeScheduledTask(mensagemModel);
 		if (mailService.deleteById(id)) {
+			if(mailService.findAll().isEmpty()){
+				agendamentoService.removeScheduledTask();
+			}
 			return ResponseType.SUCCESS_DELETE.getMessage();
 		}
 		return ResponseType.FAIL_DELETE.getMessage();
@@ -148,17 +148,11 @@ public class MensagemAdmController {
 	@ResponseStatus(code = HttpStatus.CREATED)
 	public String save(@RequestBody @Valid MensagemDTO mensagemDTO) {
 		MensagemModel mensagemModel = mapper.map(mensagemDTO, MensagemModel.class);
+		mensagemModel.setDataEnviada(mensagemModel.getDataEnvio());
+		if(mailService.findAll().isEmpty()){
+			agendamentoService.setScheduleATask(agendamentoServiceImpl);
+		}
 		mailService.save(mensagemModel);
-		if (mailService.findAll().size() == 1 && mensagemModel.getEmail() == null) {
-			mensagemModel.setFrequente(true);
-			mensagemModel.setAnual(true);
-			mailService.setEmailAnualCadastro(mailServiceImpl, mensagemModel);
-		}
-		if (mailService.findAll().size() > 1 || (mailService.findAll().size() == 1 && mensagemModel.getEmail() != null)) {
-			mailService.setScheduleATask(mailServiceImpl, mensagemDTO.getDataEnvio(), mensagemDTO.getFrequente(),
-					mensagemDTO.getAnual(), mensagemModel);
-		}
-
 		return ResponseType.SUCCESS_UPDATE.getMessage();
 	}
