+++ backend/meuegresso/src/main/java/labes/facomp/ufpa/br/meuegresso/model/MensagemModel.java
@@ -35,6 +35,9 @@ public class MensagemModel {
     @Column(name = "data_mensagem", unique = false, nullable = false)
     private LocalDateTime dataEnvio;
 
+    @Column(name = "data_enviada_mensagem", unique = false, nullable = false)
+    private LocalDateTime dataEnviada;
+
     @Builder.Default
     @Column(name = "frequente_mensagem", unique = false, nullable = true)
     private Boolean frequente = false;
